+++ CHANGELOG.md
@@ -18,6 +18,8 @@ Main (unreleased)
 
 - Add `otelcol.exporter.splunkhec` allowing to export otel data to Splunk HEC (@adlotsof)
 
+- Add `otelcol.receiver.solace` component to receive traces from a Solace broker. (@wildum)
+
 ### Bugfixes
 
 - Fixed an issue in the `prometheus.exporter.postgres` component that would leak goroutines when the target was not reachable (@dehaansa)